backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.35.0
pydantic-settings>=2.9.0
mcp>=1.26.0
fastmcp>=3.1.0
sqlalchemy[asyncio]>=2.0.40
asyncpg>=0.30.0
alembic>=1.15.0
qdrant-client>=1.14.0
celery[redis]>=5.5.0
redis>=5.3.0
torch>=2.7.0
torchaudio>=2.7.0
librosa>=0.10.2
soundfile>=0.13.0
transformers>=4.51.0
sentence-transformers>=4.1.0
lyricsgenius>=3.0.1
anthropic>=0.52.0
httpx>=0.28.0
numpy>=2.2.0
python-dotenv>=1.1.0
yt-dlp>=2024.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
